feat: move icons to shared folder
