+++ packages/react/src/icons/CloseIcon.tsx
