+++ packages/react/src/icons/FuelWalletDevelopmentIcon.tsx
