+++ packages/react/src/icons/FuelWalletIcon.tsx
