+++ packages/react/src/ui/Connect/components/ConnectorIcon.tsx
@@ -1,9 +1,9 @@
 import type { ConnectorMetadata } from 'fuels';
 
+import { FuelWalletDevelopmentIcon } from '../../../icons/FuelWalletDevelopmentIcon';
+import { FuelWalletIcon } from '../../../icons/FuelWalletIcon';
+import { FueletIcon } from '../../../icons/FueletIcon';
 import type { SvgIconProps } from '../../types';
-import { FuelWalletDevelopmentIcon } from '../icons/FuelWalletDevelopmentIcon';
-import { FuelWalletIcon } from '../icons/FuelWalletIcon';
-import { FueletIcon } from '../icons/FueletIcon';
 import { getImageUrl } from '../utils/getImageUrl';
 
 type ConnectorIconProps = {
